+++ Runtime/Behaviours/CinemachineStoryboard.cs
@@ -21,6 +21,7 @@ namespace Cinemachine
 #else
     [ExecuteInEditMode]
 #endif
+    [DisallowMultipleComponent]
     [HelpURL(Documentation.BaseURL + "manual/CinemachineStoryboard.html")]
     public class CinemachineStoryboard : CinemachineExtension
     {